16_changes on the barcode and deleting process

+++ database/migrations/2025_01_06_100744_create_products_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->text('description')->nullable();
             $table->decimal('buying_price');
             $table->decimal('selling_price');
-            $table->decimal('stock');
+            $table->integer('stock');
             $table->string('image')->nullable();
             $table->string('buying_date')->nullable();
